+++ RIN.WebAPI/Models/ClientApi/Army/UpdateArmyRanksOrderReq.cs
@@ -4,5 +4,5 @@ namespace RIN.WebAPI.Models.ClientApi;
 
 public class UpdateArmyRanksOrderReq
 {
-    [Required] public required int[] army_rank_ids { get; set; }
+    [Required] public required long[] army_rank_ids { get; set; }
 }
